backend/pyproject.toml
[project]
name = "spend-mate"
version = "0.1.0"
description = ""
authors = [
    {name = "nathanmendis",email = "nathanmendis17@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "cryptography (>=46.0.6,<47.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
